requirements.txt
celery==4.3.0
django-celery==3.3.1
requests==2.22.0
redis==3.2.0
bs4==0.0.1
celery==4.3.0
beautifulsoup4==4.8.0
